src/main/java/io/ljunggren/unit/converter/temperature/unit/si/Celsius.java
package io.ljunggren.unit.converter.temperature.unit.si;

import java.util.function.Function;

import io.ljunggren.unit.converter.System;
import io.ljunggren.unit.converter.temperature.Temperature;

public class Celsius implements Temperature {

    @Override
    public Function<Double, Double> toBaseUnit() {
        return x -> x;
    }

    @Override
    public Function<Double, Double> fromBaseUnit() {
        return x -> x;
    }

    @Override
    public System system() {
        return System.SI;
    }

    @Override
    public String label() {
        return "C";
    }

}